requirements.txt
Flask
keras
numpy
Pillow
tensorflow
scikit-image
